All classwork

+++ Summatives/M2-summative/DVDLibrary/src/main/java/com/curleyj/dvdlibrary/controller/DvDLibraryController.java
@@ -9,8 +9,11 @@ import com.curleyj.dvdlibrary.dao.DvDLibraryDao;
 import com.curleyj.dvdlibrary.dao.DvDLibraryDaoException;
 import com.curleyj.dvdlibrary.dto.dvd;
 import com.curleyj.dvdlibrary.ui.DvDLibraryView;
+import java.util.HashMap;
 
 import java.util.List;
+import java.util.Map;
+import java.util.OptionalLong;
 
 /**
  *
@@ -27,7 +30,7 @@ public class DvDLibraryController {
     
     public void run() {
         boolean keepGoing = true;
-        int menuSelection = 0;
+        String menuSelection = "";
         
         try {
             while (keepGoing) {
@@ -35,24 +38,38 @@ public class DvDLibraryController {
                 menuSelection = getMenuSelection();
 
                 switch (menuSelection) {
-                    case 1:
+                    case "1":
                         addDvd();
                         break;
-                    case 2:
+                    case "2":
                         removeDvd();
                         break;
-                    case 3:
+                    case "3":
                         editDvd();
                         break;
-                    case 4:
+                    case "4":
                         listAllDvds();
                         break;
-                    case 5:
+                    case "5":
                         listDvdInfo();
                         break;
-                    case 6:
-                        keepGoing = false;
+                    case "6":
+                        getDvdsLastNYears();
                         break;
+                    case "7":
+                        getDvdsByRating();
+                        break;
+                    case "8":
+                        getDvdsByDirector();
+                        break;
+                    case "9":
+                        getDvdByStudio();
+                        break;
+                    case "10":
+                        getAverageDvdAge();
+                        break;
+                    case "11":
+                        getNewestDvd();
                     default:
                         unknownCommand();
                 }
@@ -64,7 +81,7 @@ public class DvDLibraryController {
         }
     }
     
-    private int getMenuSelection() {
+    private String getMenuSelection() {
         return view.printMenuAndGetSelection();
     }
     
@@ -76,8 +93,20 @@ public class DvDLibraryController {
     }
     
     private void listDvdInfo() throws DvDLibraryDaoException {
+        boolean titleTest = true;
         view.displayDvdInfoBanner();
-        String title = view.dvdTitle();
+        String title = "";
+        while(titleTest) {
+            title = view.dvdTitle();
+            titleTest = dao.titleCheck(title);
+            if (!titleTest) {
+                view.displayErrorMessage("That is not a title.  Please re-input now.");
+                titleTest = true;
+            }
+            else {
+                break;
+            }
+        }
         dvd currentDvd = dao.listDvdInfo(title);
         view.listDvdInfo(currentDvd);
     }
@@ -89,17 +118,43 @@ public class DvDLibraryController {
     }
     
     private void removeDvd() throws DvDLibraryDaoException {
+        boolean titleTest = true;
+        String title = "";
         view.displayRemoveDvdBanner();
-        String currentDvd = view.dvdTitle();
-        dao.removeDvd(currentDvd);
+        while(titleTest) {
+            String currentDvd = view.dvdTitle();
+            titleTest = dao.titleCheck(currentDvd);
+            if (!titleTest) {
+                view.displayErrorMessage("That is not a title.  Please re-input now.");
+                titleTest = true;
+            }
+            else {
+                dao.removeDvd(currentDvd);
+                break;
+            }
+        }
+        
         view.displayRemoveSuccessBanner();
     }
     
     private dvd editDvd() throws DvDLibraryDaoException {
+        boolean titleTest = true;
         view.displayEditDvdBanner();
-        String currentDvd = view.dvdTitle();
-        int editChoice = view.editDvdChoice();
-        String edit = view.editDvd();
+        String currentDvd = "";
+        
+        while(titleTest) {
+            currentDvd = view.dvdTitle();
+            titleTest = dao.titleCheck(currentDvd);
+            if (!titleTest) {
+                view.displayErrorMessage("That is not a title.  Please re-input now.");
+                titleTest = true;
+            }
+            else {
+                titleTest = false;
+            }
+        }
+        String editChoice = view.editDvdChoice();
+        String edit = view.editDvd(editChoice);
         dvd editDvd = dao.editDvd(currentDvd, editChoice, edit);
         view.displayEditDvDSuccessBanner();
         return editDvd;
@@ -112,4 +167,43 @@ public class DvDLibraryController {
     private void exitMessage() {
         view.displayExitBanner();
     }
+    
+    private void getDvdsLastNYears() throws DvDLibraryDaoException {
+       view.displaylistDvdsLastNYears();
+       String sYear = view.listDvdsLastNYears();
+       int year = Integer.parseInt(sYear);
+       dao.getDvdLastNYears(year);
+       List<dvd> newList = dao.getDvdLastNYears(year);
+       view.displaySuccessDvdsLastNYears(year, newList); 
+    }
+    
+    private void getDvdsByRating() throws DvDLibraryDaoException {
+        String rating = view.dvdsByRating();
+        List<dvd> newList = dao.getDvdByRating(rating);
+        view.displaySuccessDvdByRating(rating, newList);
+    }
+    
+    private void getDvdsByDirector() throws DvDLibraryDaoException {
+        String director = view.dvdsByDirector();
+        Map<String, List<dvd>> dvdMap = dao.getDvdByDirector(director);
+        view.displaySuccessDvdByDirector(director, dvdMap);
+    }
+    
+    private void getDvdByStudio() throws DvDLibraryDaoException {
+        String studio = view.dvdsByStudio();
+        List<dvd> newList = dao.getDvdByStudio(studio);
+        view.displayDvdsByStudioSuccess(studio, newList);
+    }
+    
+    private void getAverageDvdAge() throws DvDLibraryDaoException {
+        Double age = dao.getAverageDvdAge();
+        view.displayAverageAge(age);
+    }
+    
+    private void getNewestDvd() throws DvDLibraryDaoException {
+       String dvdTitle =  dao.getNewestDvd();
+       
+       view.getNewestDvd(dvdTitle);
+    }
+    
 }
